Instructions for adding toolchain & getting data
Added instructions for adding new toolchains, and getting sources and
binaries.

Also refactored setup.py to make it easier to add new toolchains.
Closes #45.
