+++ tuscan/utilities.py
@@ -32,11 +32,37 @@ from json import dumps
 from random import random, seed
 from re import search
 from shutil import move
-from subprocess import run, PIPE, Popen, STDOUT, TimeoutExpired
+from subprocess import run, PIPE, Popen, STDOUT, TimeoutExpired, DEVNULL
 from sys import stderr
 from tempfile import NamedTemporaryFile as tempfile
 from textwrap import dedent
 from time import sleep
+
+
+def run_cmd(cmd, as_root=False, output=True):
+    time = timestamp()
+
+    if output:
+        cmd_out=PIPE
+    else:
+        cmd_out=DEVNULL
+
+    if not as_root:
+        cmd = "sudo -u tuscan " + cmd
+
+    cp = run(cmd.split(), stdout=cmd_out, stderr=STDOUT,
+             universal_newlines=True)
+
+    if output:
+        lines = cp.stdout.splitlines()
+    else:
+        lines = []
+
+    if cp.returncode:
+        log("die", cmd, lines, time)
+        exit(1)
+    else:
+        log("command", cmd, lines, time)
 
 
 def interpret_bash_array(pkgbuild, array_name):
